requirements.txt
streamlit
streamlit_option_menu
pillow
fastbook
pathlib

#for api
fastapi
uvicorn[standard]
python-multipart

#for google drive
#gdown
requests